FoldsCUDA.jl

FoldsCUDA.FoldsCUDA — Module

FoldsCUDA

FoldsCUDA.jl provides Transducers.jl-compatible fold (reduce) implemented using CUDA.jl. This brings the transducers and reducing function combinators implemented in Transducers.jl to GPU. Furthermore, using FLoops.jl, you can write parallel for loops that run on GPU.

API

foldx_cuda: a GPU equivalent of parallel extended fold Transducers.foldxt.

CUDAEx: a parallel loop executor for FLoops.jl.

See the documentation of Transducers.jl and FLoops.jl for more information.

Examples

findmax using FLoops.jl

You can pass CUDA executor FoldsCUDA.CUDAEx() to @floop to run a parallel for loop on GPU:

julia> using FoldsCUDA, CUDA, FLoops

julia> using GPUArrays: @allowscalar

julia> xs = CUDA.rand(10^8);

julia> @allowscalar xs[100] = 2;

julia> @allowscalar xs[200] = 2;

julia> @floop CUDAEx() for (x, i) in zip(xs, eachindex(xs))
           @reduce() do (imax = -1; i), (xmax = -Inf32; x)
               if xmax < x
                   xmax = x
                   imax = i
               end
           end
       end

julia> xmax
2.0f0

julia> imax  # the *first* position for the largest value
100

extrema using Transducers.TeeRF

julia> using Transducers

julia> @allowscalar xs[300] = -0.5;

julia> foldx_cuda(TeeRF(min, max), xs)
(-0.5f0, 2.0f0)

julia> foldx_cuda(TeeRF(min, max), (2x for x in xs))  # iterator comprehension works
(-1.0f0, 4.0f0)

julia> foldx_cuda(TeeRF(min, max), Map(x -> 2x), xs)  # equivalent, using a transducer
(-1.0f0, 4.0f0)

More examples

For more examples, see the examples section in the documentation.

FoldsCUDA.foldx_cuda — Function

foldx_cuda(op[, xf], xs; init)
transduce_cuda(op[, xf], init, xs)

Extended fold backed up by CUDA.

FoldsCUDA.transduce_cuda — Function

foldx_cuda(op[, xf], xs; init)
transduce_cuda(op[, xf], init, xs)

Extended fold backed up by CUDA.

FoldsCUDA.CUDAEx — Type

CUDAEx()

FLoops.jl executor implemented using CUDA.jl.